+++ watchme/utils/__init__.py
@@ -9,6 +9,7 @@ from .terminal import (
 from .fileio import (
     copyfile,
     get_userhome,
+    get_user,
     generate_temporary_file,
     mkdir_p,
     get_tmpdir,